Avoid throwing errors while requiring files for coverage.
To ensure coverage percentages are accurate we added some code to automatically require all modules in the application/addon.  If an error was detected while loading the module we annotated it with information on why it occurred, but still rethrew the error.

In retrospect, it seems that a number of applications have dormant (and invalid) modules that exist in the app (at least during tests).  Our prior concept/assertion that rethrowing an error was the correct path turns out to itself be painful.

This changes that error into a warning (because I truly believe we should encourage folks to fix these scenarios).