fix `InvalidStateError` exception
This commit fixes #244.

The xhr.responseText property cannot be accessed
unless xhr.responseType is either 'text' or '',
otherwise it throws an error:

Uncaught InvalidStateError: Failed to read the 'responseText' property from 'XMLHttpRequest':
The value is only accessible if the object's 'responseType' is '' or 'text' (was 'json')

This will therefore happen in handleCoverageResponse whenever
a request either rejects or returns falsy.

This commit introduces a check to only read from the 'responseText'
property in a valid circumstance.