Use `require.resolve` instead of mutating required module.
The code in `requireBabelPlugin` was attempting to ensure that the cache
key used by broccoli-babel-transpiler incorporated the
`babel-plugin-istanbul` directory (so that the cache would be
invalidated if it was modified / reinstalled).

At the time, this was the best we could do, but
broccoli-babel-transpiler now properly caches plugins when provided an
absolute path (which is what `require.resolve` returns). This _also_
means that consumers using ember-cli-code-coverage will now be able to
have Babel parralellization (which is not possible with the prior
implementation).